Add basic useState implementation.
Hooks can only be used in functions. React.useState can be used
multiple times in a function and set to either a function or a value.
useState returns an array of the value, and method to change the value.
THe method to change the value can simplistically change the state of one
or more properties of an object, depending on what useState sets initially.
Object should have their properties deconstructed '{}' in order to use values.